multiple unit test changes and significant changes made to handling of XML
