+++ src/functional/destroyMOEW.ts
@@ -4,25 +4,23 @@ import { outsideOfPowerChartError } from '../utils';
 /**
  * Destroys the modal order entry window (MOEW).
  * @param {number} moewHandle - the handle to the MOEW.
- * @returns a `Promise` which resolves to null. Null appears to be returned upon either a successful or unsuccessful destruction.
+ * @description The value returned by DestroyMOEW() is not used and is not believed to be meaningful, but is logged to the console for development purposes.
+ * @returns a `Promise` indicating whether called from PowerChart or not.
  * @throws `Error` if an unexpected error occurs.
  */
 
 export async function destroyMOEWAsync(
   moewHandle: number
-): Promise<DestroyMOEWReturn> {
-  let retData: DestroyMOEWReturn = {
+): Promise<PowerChartReturn> {
+  let retData: PowerChartReturn = {
     inPowerChart: true,
-    retVal: null,
   };
 
   // Create the DiscernObjectFactory and use that to call DestroyMOEW() with the handle from above
   try {
     const dcof = await window.external.DiscernObjectFactory('POWERORDERS');
     const response = await dcof.DestroyMOEW(moewHandle);
-
-    // Set the retValue equal to the return (which appears to always be null)
-    retData.retVal === response;
+    console.log('Response from DestroyMOEW() is: ', response);
   } catch (e) {
     //If outside of PowerChart, set the output to reflect that
     if (outsideOfPowerChartError(e)) {
@@ -36,7 +34,3 @@ export async function destroyMOEWAsync(
   // Return the retData object when complete
   return retData;
 }
-
-export type DestroyMOEWReturn = PowerChartReturn & {
-  retVal: null;
-};
